+++ SustitucionMOAModel/Entities/LogTable.cs
@@ -1,18 +1,14 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace SustitucionMOAModel.Entities
 {
     [Table("LogTable", Schema = "logs")]
     public class LogTable
     {
-        [Key]        
-        public int Id { get; set; }
+        [Key]
+        public long Id { get; set; }
         public DateTime Date { get; set; }
         public string Level { get; set; }
         public string Logger { get; set; }
